requirements.txt
requests==2.31.0
groq==1.1.2
python-dotenv==1.0.1
